+++ app/helpers/comparisons_helper.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ComparisonsHelper
+  def checked?(slug)
+    session[:comparison]&.include?(slug)
+  end
+end